register_signing_box for rust
