+++ ton_client/src/json_interface/mod.rs
@@ -12,6 +12,7 @@
  *
  */
 
+pub(crate) mod crypto;
 pub(crate) mod handlers;
 pub(crate) mod interop;
 pub(crate) mod net;
